lavc/ccaption_dec: reap_screen after flipping on EOC
The EOC command stands for "end of caption" aka "display buffer". It's
used with POPON mode, where characters are written to an off-screen
buffer and EOC flips the buffers to display what has been written so
far. Thus, it makes sense to reap the screen *after* flipping the active
screen, not *before*.

The previous behavior was simply wrong, but masked by other bugs also
fixed in this patchset.