lavc/ccaption_dec: implement musical glyph
This is the most commonly used character from the special north-american
character set. All the non-standard charsets are "optional" according to
the spec, and we currently implement none of them. This commit adds
support for "♪" which very popular (and has no ascii fallback), typically
used to indicate lyrics in captions.