avformat/avio: add new_http_request callback
This callback is implemented on AVIOContext which are backed by an
HTTP URLContext. This allows the owner of the IO context to both
discover that it is capable of making additional http requests, and
to make such requests without including http.h. In particular this
avoids the use of internal functions to fetch the URLContext, and
prevents accidentally calling ff_http_do_new_request on a URLContext
which is not actually part of an HTTPContext.

Signed-off-by: Aman Gupta <aman@tmm1.net>
Signed-off-by: wm4 <nfxjfg@googlemail.com>