avcodec/videotoolbox: fix decoding of some hevc videos
See https://s3.amazonaws.com/tmm1/videotoolbox/germany-hevc-zdf.ts

This moves the hw_frames_ctx reference from the AVFrame to a new
VTHWFrame which now holds both the CVPixelBufferRef and the AVBuffer.

Previously we would set the hw_frames_context on the AVFrame directly,
but on some videos the HEVC decoder would copy and move frames around
to reorder them, and in the process could accidentally free the
hw_frames_ctx, such that it was NULL by the time videotoolbox_postproc_frame()
got called. This appears to be a bug specific to the HEVC decoder, but
after several days of debugging I was unable to figure out where
the memory management issue is occurring.

This patch works around the issue by assigning hw_frames_ctx only once
the AVFrame is about to returned to the user. Fixes playback of the
HEVC sample linked above on both macOS and iOS.

Signed-off-by: Aman Gupta <aman@tmm1.net>