avformat/http: parse Content-Range responses with unknown document size
The HTTP spec allows '*' for document size in Content-Range response headers,
to specify that the file size is unknown:

    Content-Range: <unit> <range-start>-<range-end>/<size>
    Content-Range: <unit> <range-start>-<range-end>/*
    Content-Range: <unit> */<size>

In the case where the filesize is unknown, this patch keeps track of a new fileend variable
which is the approximate size based on the range and data returned by the server.

Practically speaking, this patch can be used effectively "tail" a stream that is
still being appended to. For example, a PVR system may record an mpegts stream to disk,
and serve up the same file to players while the stream is actively being recorded.

By adding a `Content-Range: bytes X-Y/*` header to the response, the server can
inform the ffmpeg-based player that the stream is "active" and that the file size
will continue to grow. For every incoming request, the server would:

- open file handle to recording in progress
- seek file handle to offset requested by the incoming `Range` header
- query the current file size of the file
- return a `Content-Range` header with the format `bytes <start offset>-<current file size>/*`
- copy bytes from the file to the http socket
- on eof, wait and try reading from the file again to "tail" it as long as the recording is "active"

On the ffmpeg size, we simply use MAX(s->off, s->fileend) to decide how large the file is
whenever requested via AVSEEK_SIZE. This effectively allows ffmpeg based players to
watch and seek around a growing mpegts stream even while it is being appended to.

Signed-off-by: Aman Karmani <aman@tmm1.net>